requirements.txt
astroid==2.1.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1
django-appconf==1.0.2
django-debug-toolbar==1.10.1
django-imagekit==4.0.2
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.6
mccabe==0.6.1
pilkit==2.0
Pillow==5.4.1
psycopg2-binary==2.7.6.1
pylint==2.2.2
python-decouple==3.1
pytz==2018.9
six==1.12.0
sqlparse==0.2.4
Unipath==1.0
whitenoise==4.1.2
wrapt==1.11.0
